always lower all classes to avoid potential bugs
Previously "lowerClass" tried to early-out if it could detect that class lowering didn't need to be applied. This probably isn't a good idea because it could potentially introduce correctness bugs if the code is changed without changing the check. Also the check is already pretty complex and hard to reason about. It seems best to just always apply the transform and have the transform be a no-op if it turns out it wasn't needed. This shouldn't be a big performance hit because it only iterates over properties on the class body.